+++ UI/Services/Properties/AssemblyInfo.cs
@@ -5,4 +5,4 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyTitle("Macabre2D.UI.Services")]
 [assembly: AssemblyDescription("Services for the Macabre2D UI.")]
 [assembly: Guid("1571ed27-bbde-4b17-b80b-0da350770ac7")]
-[assembly: InternalsVisibleTo("Macabre2D.Tests")]
+[assembly: InternalsVisibleTo("Macabre2D.UI.Tests")]